feat: add listing data to rental detail skeleton
